RuboCop: Performance/ChainArrayAllocation
Removing the grouping from `PARAMETERS_REGEXP_PATTERN` results in a flat
array coming from the call to `scan`. The resulting matches are _nearly_
the same with the exception of whitespace (thus the new call to `strip`
in the `map` call).

This addresses the chain array allocation warning and simplifies the
code _a little bit_.

tbh I'm not sure why I'm using a negative lookbehind in the regular
expression, but I'll have to trust myself from a couple years ago.

:grimace: